Use scrt-utils and move tests
