+++ scrt-snip20/src/state.rs
@@ -8,10 +8,11 @@ use secret_toolkit::storage::{TypedStore, TypedStoreMut};
 
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use serde::de::DeserializeOwned;
+
+use cosmwasm_utils::viewing_key::ViewingKey;
 
 use crate::msg::{status_level_to_u8, u8_to_status_level, ContractStatusLevel};
-use crate::viewing_key::ViewingKey;
-use serde::de::DeserializeOwned;
 
 pub static CONFIG_KEY: &[u8] = b"config";
 pub const PREFIX_TXS: &[u8] = b"transfers";
